+++ app/lib/auto-launch.js
@@ -78,6 +78,11 @@ class BalloonAutoLaunch {
           });
         }
       }).catch(err => {
+        if(err.message && err.message.match(/(-1743)/)) {
+          logger.error('Autolaounch can\'t be controlled by balloon as it is not allowed to send Apple-Events to System Events.', {category: 'autolaunch'});
+          return resolve();
+        }
+
         logger.error('could not query autolaunch state', {category: 'autolaunch', err});
         return reject(err);
       });
@@ -104,6 +109,11 @@ class BalloonAutoLaunch {
           });
         }
       }).catch(err => {
+        if(err.message && err.message.match(/(-1743)/)) {
+          logger.error('Autolaounch can\'t be controlled by balloon as it is not allowed to send Apple-Events to System Events.', {category: 'autolaunch'});
+          return resolve();
+        }
+
         logger.error('could not query autolaunch state', {category: 'autolaunch', err});
         return reject(err);
       });